Test backend

+++ internal/bowls/service.go
@@ -8,12 +8,11 @@ func NewBowlsService(repo BowlRepository) *BowlsService {
 	return &BowlsService{repo}
 }
 
-
 func (s *BowlsService) GetAll() ([]Bowl, error) {
 	return s.repo.GetAll()
 }
 
-func (s *BowlsService) GetById(id int) (*Bowl, error) {
+func (s *BowlsService) GetById(id int64) (*Bowl, error) {
 	return s.repo.GetById(id)
 }
 
@@ -21,10 +20,10 @@ func (s *BowlsService) Create(name string, description *string) (*Bowl, error) {
 	return s.repo.Create(name, description)
 }
 
-func (s *BowlsService) UpdateById(id int, update BowlUpdate) (*Bowl, error) {
+func (s *BowlsService) UpdateById(id int64, update BowlUpdate) (*Bowl, error) {
 	return s.repo.UpdateById(id, update)
 }
 
-func (s *BowlsService) DeleteById(id int) error {
+func (s *BowlsService) DeleteById(id int64) error {
 	return s.repo.DeleteById(id)
 }
